librados: remove deprecated object listing interfaces
These have been deprecated since giant.
