+++ src/test/librados/list.cc
@@ -22,26 +22,22 @@ typedef RadosTestECNS LibRadosListEC;
 typedef RadosTestECPPNS LibRadosListECPP;
 typedef RadosTestNP LibRadosListNP;
 
+
 TEST_F(LibRadosList, ListObjects) {
   char buf[128];
   memset(buf, 0xcc, sizeof(buf));
   ASSERT_EQ(0, rados_write(ioctx, "foo", buf, sizeof(buf), 0));
   rados_list_ctx_t ctx;
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
   const char *entry;
   bool foundit = false;
-  while (rados_objects_list_next(ctx, &entry, NULL) != -ENOENT) {
+  while (rados_nobjects_list_next(ctx, &entry, NULL, NULL) != -ENOENT) {
     foundit = true;
     ASSERT_EQ(std::string(entry), "foo");
   }
   ASSERT_TRUE(foundit);
-  rados_objects_list_close(ctx);
+  rados_nobjects_list_close(ctx);
 }
-
-
-#pragma GCC diagnostic ignored "-Wpragmas"
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
 
 TEST_F(LibRadosListPP, ListObjectsPP) {
   char buf[128];
@@ -49,11 +45,11 @@ TEST_F(LibRadosListPP, ListObjectsPP) {
   bufferlist bl1;
   bl1.append(buf, sizeof(buf));
   ASSERT_EQ(0, ioctx.write("foo", bl1, sizeof(buf), 0));
-  ObjectIterator iter(ioctx.objects_begin());
+  NObjectIterator iter(ioctx.nobjects_begin());
   bool foundit = false;
-  while (iter != ioctx.objects_end()) {
+  while (iter != ioctx.nobjects_end()) {
     foundit = true;
-    ASSERT_EQ((*iter).first, "foo");
+    ASSERT_EQ((*iter).get_oid(), "foo");
     ++iter;
   }
   ASSERT_TRUE(foundit);
@@ -65,21 +61,21 @@ TEST_F(LibRadosListPP, ListObjectsTwicePP) {
   bufferlist bl1;
   bl1.append(buf, sizeof(buf));
   ASSERT_EQ(0, ioctx.write("foo", bl1, sizeof(buf), 0));
-  ObjectIterator iter(ioctx.objects_begin());
+  NObjectIterator iter(ioctx.nobjects_begin());
   bool foundit = false;
-  while (iter != ioctx.objects_end()) {
+  while (iter != ioctx.nobjects_end()) {
     foundit = true;
-    ASSERT_EQ((*iter).first, "foo");
+    ASSERT_EQ((*iter).get_oid(), "foo");
     ++iter;
   }
   ASSERT_TRUE(foundit);
   ++iter;
-  ASSERT_TRUE(iter == ioctx.objects_end());
+  ASSERT_TRUE(iter == ioctx.nobjects_end());
   foundit = false;
   iter.seek(0);
-  while (iter != ioctx.objects_end()) {
+  while (iter != ioctx.nobjects_end()) {
     foundit = true;
-    ASSERT_EQ((*iter).first, "foo");
+    ASSERT_EQ((*iter).get_oid(), "foo");
     ++iter;
   }
   ASSERT_TRUE(foundit);
@@ -93,28 +89,28 @@ TEST_F(LibRadosListPP, ListObjectsCopyIterPP) {
   ASSERT_EQ(0, ioctx.write("foo", bl1, sizeof(buf), 0));
 
   // make sure this is still valid after the original iterators are gone
-  ObjectIterator iter3;
+  NObjectIterator iter3;
   {
-    ObjectIterator iter(ioctx.objects_begin());
-    ObjectIterator iter2(iter);
+    NObjectIterator iter(ioctx.nobjects_begin());
+    NObjectIterator iter2(iter);
     iter3 = iter2;
-    ASSERT_EQ((*iter).first, "foo");
+    ASSERT_EQ((*iter).get_oid(), "foo");
     ++iter;
-    ASSERT_TRUE(iter == ioctx.objects_end());
+    ASSERT_TRUE(iter == ioctx.nobjects_end());
     ++iter;
-    ASSERT_TRUE(iter == ioctx.objects_end());
+    ASSERT_TRUE(iter == ioctx.nobjects_end());
 
-    ASSERT_EQ(iter2->first, "foo");
-    ASSERT_EQ(iter3->first, "foo");
+    ASSERT_EQ(iter2->get_oid(), "foo");
+    ASSERT_EQ(iter3->get_oid(), "foo");
     ++iter2;
-    ASSERT_TRUE(iter2 == ioctx.objects_end());
+    ASSERT_TRUE(iter2 == ioctx.nobjects_end());
   }
 
-  ASSERT_EQ(iter3->first, "foo");
+  ASSERT_EQ(iter3->get_oid(), "foo");
   iter3 = iter3;
-  ASSERT_EQ(iter3->first, "foo");
+  ASSERT_EQ(iter3->get_oid(), "foo");
   ++iter3;
-  ASSERT_TRUE(iter3 == ioctx.objects_end());
+  ASSERT_TRUE(iter3 == ioctx.nobjects_end());
 }
 
 TEST_F(LibRadosListPP, ListObjectsEndIter) {
@@ -124,27 +120,27 @@ TEST_F(LibRadosListPP, ListObjectsEndIter) {
   bl1.append(buf, sizeof(buf));
   ASSERT_EQ(0, ioctx.write("foo", bl1, sizeof(buf), 0));
 
-  ObjectIterator iter(ioctx.objects_begin());
-  ObjectIterator iter_end(ioctx.objects_end());
-  ObjectIterator iter_end2 = ioctx.objects_end();
+  NObjectIterator iter(ioctx.nobjects_begin());
+  NObjectIterator iter_end(ioctx.nobjects_end());
+  NObjectIterator iter_end2 = ioctx.nobjects_end();
   ASSERT_TRUE(iter_end == iter_end2);
-  ASSERT_TRUE(iter_end == ioctx.objects_end());
-  ASSERT_TRUE(iter_end2 == ioctx.objects_end());
+  ASSERT_TRUE(iter_end == ioctx.nobjects_end());
+  ASSERT_TRUE(iter_end2 == ioctx.nobjects_end());
 
-  ASSERT_EQ(iter->first, "foo");
+  ASSERT_EQ(iter->get_oid(), "foo");
   ++iter;
-  ASSERT_TRUE(iter == ioctx.objects_end());
+  ASSERT_TRUE(iter == ioctx.nobjects_end());
   ASSERT_TRUE(iter == iter_end);
   ASSERT_TRUE(iter == iter_end2);
-  ObjectIterator iter2 = iter;
-  ASSERT_TRUE(iter2 == ioctx.objects_end());
+  NObjectIterator iter2 = iter;
+  ASSERT_TRUE(iter2 == ioctx.nobjects_end());
   ASSERT_TRUE(iter2 == iter_end);
   ASSERT_TRUE(iter2 == iter_end2);
 }
 
-static void check_list(std::set<std::string>& myset, rados_list_ctx_t& ctx)
+static void check_list(std::set<std::string>& myset, rados_list_ctx_t& ctx, std::string check_nspace)
 {
-  const char *entry;
+  const char *entry, *nspace;
   std::set<std::string> orig_set(myset);
   /**
    * During splitting, we might see duplicate items.
@@ -152,15 +148,22 @@ static void check_list(std::set<std::string>& myset, rados_list_ctx_t& ctx)
    * we don't hit ENOENT until we have hit every item in myset
    * at least once.
    */
-  while (rados_objects_list_next(ctx, &entry, NULL) != -ENOENT) {
-    ASSERT_TRUE(orig_set.end() != orig_set.find(std::string(entry)));
-    myset.erase(std::string(entry));
+  int ret;
+  while ((ret = rados_nobjects_list_next(ctx, &entry, NULL, &nspace)) == 0) {
+    std::string test_name;
+    if (check_nspace == all_nspaces) {
+      test_name = std::string(nspace) + ":" + std::string(entry);
+    } else {
+      ASSERT_TRUE(std::string(nspace) == check_nspace);
+      test_name = std::string(entry);
+    }
+
+    ASSERT_TRUE(orig_set.end() != orig_set.find(test_name));
+    myset.erase(test_name);
   }
+  ASSERT_EQ(-ENOENT, ret);
   ASSERT_TRUE(myset.empty());
 }
-
-#pragma GCC diagnostic pop
-#pragma GCC diagnostic warning "-Wpragmas"
 
 TEST_F(LibRadosList, ListObjectsNS) {
   char buf[128];
@@ -180,7 +183,7 @@ TEST_F(LibRadosList, ListObjectsNS) {
   ASSERT_EQ(0, rados_write(ioctx, "foo6", buf, sizeof(buf), 0));
   ASSERT_EQ(0, rados_write(ioctx, "foo7", buf, sizeof(buf), 0));
 
-  std::set<std::string> def, ns1, ns2;
+  std::set<std::string> def, ns1, ns2, all;
   def.insert(std::string("foo1"));
   def.insert(std::string("foo2"));
   def.insert(std::string("foo3"));
@@ -189,38 +192,44 @@ TEST_F(LibRadosList, ListObjectsNS) {
   ns1.insert(std::string("foo5"));
   ns2.insert(std::string("foo6"));
   ns2.insert(std::string("foo7"));
+  all.insert(std::string(":foo1"));
+  all.insert(std::string(":foo2"));
+  all.insert(std::string(":foo3"));
+  all.insert(std::string("ns1:foo1"));
+  all.insert(std::string("ns1:foo4"));
+  all.insert(std::string("ns1:foo5"));
+  all.insert(std::string("ns2:foo6"));
+  all.insert(std::string("ns2:foo7"));
 
   rados_list_ctx_t ctx;
   // Check default namespace ""
   rados_ioctx_set_namespace(ioctx, "");
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
-  check_list(def, ctx);
-  rados_objects_list_close(ctx);
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  check_list(def, ctx, "");
+  rados_nobjects_list_close(ctx);
 
-  // Check default namespace "ns1"
+  // Check namespace "ns1"
   rados_ioctx_set_namespace(ioctx, "ns1");
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
-  check_list(ns1, ctx);
-  rados_objects_list_close(ctx);
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  check_list(ns1, ctx, "ns1");
+  rados_nobjects_list_close(ctx);
 
-  // Check default namespace "ns2"
+  // Check namespace "ns2"
   rados_ioctx_set_namespace(ioctx, "ns2");
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
-  check_list(ns2, ctx);
-  rados_objects_list_close(ctx);
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  check_list(ns2, ctx, "ns2");
+  rados_nobjects_list_close(ctx);
 
-  // Can't specify all namespaces using old interface
+  // Check ALL namespaces
   rados_ioctx_set_namespace(ioctx, LIBRADOS_ALL_NSPACES);
-  ASSERT_EQ(-EINVAL, rados_objects_list_open(ioctx, &ctx));
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  check_list(all, ctx, all_nspaces);
+  rados_nobjects_list_close(ctx);
 }
 
-#pragma GCC diagnostic ignored "-Wpragmas"
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
-
-static void check_listpp(std::set<std::string>& myset, IoCtx& ioctx)
+static void check_listpp(std::set<std::string>& myset, IoCtx& ioctx, std::string check_nspace)
 {
-  ObjectIterator iter(ioctx.objects_begin());
+  NObjectIterator iter(ioctx.nobjects_begin());
   std::set<std::string> orig_set(myset);
   /**
    * During splitting, we might see duplicate items.
@@ -228,9 +237,16 @@ static void check_listpp(std::set<std::string>& myset, IoCtx& ioctx)
    * we don't hit ENOENT until we have hit every item in myset
    * at least once.
    */
-  while (iter != ioctx.objects_end()) {
-    ASSERT_TRUE(orig_set.end() != orig_set.find(std::string((*iter).first)));
-    myset.erase(std::string((*iter).first));
+  while (iter != ioctx.nobjects_end()) {
+    std::string test_name;
+    if (check_nspace == all_nspaces) {
+      test_name = iter->get_nspace() + ":" + iter->get_oid();
+    } else {
+      ASSERT_TRUE(iter->get_nspace() == check_nspace);
+      test_name = iter->get_oid();
+    }
+    ASSERT_TRUE(orig_set.end() != orig_set.find(test_name));
+    myset.erase(test_name);
     ++iter;
   }
   ASSERT_TRUE(myset.empty());
@@ -256,7 +272,7 @@ TEST_F(LibRadosListPP, ListObjectsPPNS) {
   ASSERT_EQ(0, ioctx.write("foo6", bl1, sizeof(buf), 0));
   ASSERT_EQ(0, ioctx.write("foo7", bl1, sizeof(buf), 0));
 
-  std::set<std::string> def, ns1, ns2;
+  std::set<std::string> def, ns1, ns2, all;
   def.insert(std::string("foo1"));
   def.insert(std::string("foo2"));
   def.insert(std::string("foo3"));
@@ -265,18 +281,26 @@ TEST_F(LibRadosListPP, ListObjectsPPNS) {
   ns1.insert(std::string("foo5"));
   ns2.insert(std::string("foo6"));
   ns2.insert(std::string("foo7"));
+  all.insert(std::string(":foo1"));
+  all.insert(std::string(":foo2"));
+  all.insert(std::string(":foo3"));
+  all.insert(std::string("ns1:foo1"));
+  all.insert(std::string("ns1:foo4"));
+  all.insert(std::string("ns1:foo5"));
+  all.insert(std::string("ns2:foo6"));
+  all.insert(std::string("ns2:foo7"));
 
   ioctx.set_namespace("");
-  check_listpp(def, ioctx);
+  check_listpp(def, ioctx, "");
 
   ioctx.set_namespace("ns1");
-  check_listpp(ns1, ioctx);
+  check_listpp(ns1, ioctx, "ns1");
 
   ioctx.set_namespace("ns2");
-  check_listpp(ns2, ioctx);
+  check_listpp(ns2, ioctx, "ns2");
 
   ioctx.set_namespace(all_nspaces);
-  EXPECT_THROW(check_listpp(def, ioctx), std::runtime_error);
+  check_listpp(all, ioctx, all_nspaces);
 }
 
 TEST_F(LibRadosListPP, ListObjectsManyPP) {
@@ -289,13 +313,13 @@ TEST_F(LibRadosListPP, ListObjectsManyPP) {
     ASSERT_EQ(0, ioctx.write(stringify(i), bl, bl.length(), 0));
   }
 
-  librados::ObjectIterator it = ioctx.objects_begin();
+  librados::NObjectIterator it = ioctx.nobjects_begin();
   std::set<std::string> saw_obj;
   std::set<int> saw_pg;
-  for (; it != ioctx.objects_end(); ++it) {
-    std::cout << it->first
+  for (; it != ioctx.nobjects_end(); ++it) {
+    std::cout << it->get_oid()
 	      << " " << it.get_pg_hash_position() << std::endl;
-    saw_obj.insert(it->first);
+    saw_obj.insert(it->get_oid());
     saw_pg.insert(it.get_pg_hash_position());
   }
   std::cout << "saw " << saw_pg.size() << " pgs " << std::endl;
@@ -315,27 +339,27 @@ TEST_F(LibRadosList, ListObjectsStart) {
   }
 
   rados_list_ctx_t ctx;
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
   std::map<int, std::set<std::string> > pg_to_obj;
   const char *entry;
-  while (rados_objects_list_next(ctx, &entry, NULL) == 0) {
-    uint32_t pos = rados_objects_list_get_pg_hash_position(ctx);
+  while (rados_nobjects_list_next(ctx, &entry, NULL, NULL) == 0) {
+    uint32_t pos = rados_nobjects_list_get_pg_hash_position(ctx);
     std::cout << entry << " " << pos << std::endl;
     pg_to_obj[pos].insert(entry);
   }
-  rados_objects_list_close(ctx);
+  rados_nobjects_list_close(ctx);
 
   std::map<int, std::set<std::string> >::reverse_iterator p =
     pg_to_obj.rbegin();
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
   while (p != pg_to_obj.rend()) {
-    ASSERT_EQ((uint32_t)p->first, rados_objects_list_seek(ctx, p->first));
-    ASSERT_EQ(0, rados_objects_list_next(ctx, &entry, NULL));
+    ASSERT_EQ((uint32_t)p->first, rados_nobjects_list_seek(ctx, p->first));
+    ASSERT_EQ(0, rados_nobjects_list_next(ctx, &entry, NULL, NULL));
     std::cout << "have " << entry << " expect one of " << p->second << std::endl;
     ASSERT_TRUE(p->second.count(entry));
     ++p;
   }
-  rados_objects_list_close(ctx);
+  rados_nobjects_list_close(ctx);
 }
 
 TEST_F(LibRadosListPP, ListObjectsStartPP) {
@@ -348,20 +372,20 @@ TEST_F(LibRadosListPP, ListObjectsStartPP) {
     ASSERT_EQ(0, ioctx.write(stringify(i), bl, bl.length(), 0));
   }
 
-  librados::ObjectIterator it = ioctx.objects_begin();
+  librados::NObjectIterator it = ioctx.nobjects_begin();
   std::map<int, std::set<std::string> > pg_to_obj;
-  for (; it != ioctx.objects_end(); ++it) {
-    std::cout << it->first << " " << it.get_pg_hash_position() << std::endl;
-    pg_to_obj[it.get_pg_hash_position()].insert(it->first);
+  for (; it != ioctx.nobjects_end(); ++it) {
+    std::cout << it->get_oid() << " " << it.get_pg_hash_position() << std::endl;
+    pg_to_obj[it.get_pg_hash_position()].insert(it->get_oid());
   }
 
   std::map<int, std::set<std::string> >::reverse_iterator p =
     pg_to_obj.rbegin();
-  it = ioctx.objects_begin(p->first);
+  it = ioctx.nobjects_begin(p->first);
   while (p != pg_to_obj.rend()) {
     ASSERT_EQ((uint32_t)p->first, it.seek(p->first));
-    std::cout << "have " << it->first << " expect one of " << p->second << std::endl;
-    ASSERT_TRUE(p->second.count(it->first));
+    std::cout << "have " << it->get_oid() << " expect one of " << p->second << std::endl;
+    ASSERT_TRUE(p->second.count(it->get_oid()));
     ++p;
   }
 }
@@ -371,15 +395,15 @@ TEST_F(LibRadosListEC, ListObjects) {
   memset(buf, 0xcc, sizeof(buf));
   ASSERT_EQ(0, rados_write(ioctx, "foo", buf, sizeof(buf), 0));
   rados_list_ctx_t ctx;
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
   const char *entry;
   bool foundit = false;
-  while (rados_objects_list_next(ctx, &entry, NULL) != -ENOENT) {
+  while (rados_nobjects_list_next(ctx, &entry, NULL, NULL) != -ENOENT) {
     foundit = true;
     ASSERT_EQ(std::string(entry), "foo");
   }
   ASSERT_TRUE(foundit);
-  rados_objects_list_close(ctx);
+  rados_nobjects_list_close(ctx);
 }
 
 TEST_F(LibRadosListECPP, ListObjectsPP) {
@@ -388,11 +412,11 @@ TEST_F(LibRadosListECPP, ListObjectsPP) {
   bufferlist bl1;
   bl1.append(buf, sizeof(buf));
   ASSERT_EQ(0, ioctx.write("foo", bl1, sizeof(buf), 0));
-  ObjectIterator iter(ioctx.objects_begin());
+  NObjectIterator iter(ioctx.nobjects_begin());
   bool foundit = false;
-  while (iter != ioctx.objects_end()) {
+  while (iter != ioctx.nobjects_end()) {
     foundit = true;
-    ASSERT_EQ((*iter).first, "foo");
+    ASSERT_EQ((*iter).get_oid(), "foo");
     ++iter;
   }
   ASSERT_TRUE(foundit);
@@ -404,21 +428,21 @@ TEST_F(LibRadosListECPP, ListObjectsTwicePP) {
   bufferlist bl1;
   bl1.append(buf, sizeof(buf));
   ASSERT_EQ(0, ioctx.write("foo", bl1, sizeof(buf), 0));
-  ObjectIterator iter(ioctx.objects_begin());
+  NObjectIterator iter(ioctx.nobjects_begin());
   bool foundit = false;
-  while (iter != ioctx.objects_end()) {
+  while (iter != ioctx.nobjects_end()) {
     foundit = true;
-    ASSERT_EQ((*iter).first, "foo");
+    ASSERT_EQ((*iter).get_oid(), "foo");
     ++iter;
   }
   ASSERT_TRUE(foundit);
   ++iter;
-  ASSERT_TRUE(iter == ioctx.objects_end());
+  ASSERT_TRUE(iter == ioctx.nobjects_end());
   foundit = false;
   iter.seek(0);
-  while (iter != ioctx.objects_end()) {
+  while (iter != ioctx.nobjects_end()) {
     foundit = true;
-    ASSERT_EQ((*iter).first, "foo");
+    ASSERT_EQ((*iter).get_oid(), "foo");
     ++iter;
   }
   ASSERT_TRUE(foundit);
@@ -432,28 +456,28 @@ TEST_F(LibRadosListECPP, ListObjectsCopyIterPP) {
   ASSERT_EQ(0, ioctx.write("foo", bl1, sizeof(buf), 0));
 
   // make sure this is still valid after the original iterators are gone
-  ObjectIterator iter3;
+  NObjectIterator iter3;
   {
-    ObjectIterator iter(ioctx.objects_begin());
-    ObjectIterator iter2(iter);
+    NObjectIterator iter(ioctx.nobjects_begin());
+    NObjectIterator iter2(iter);
     iter3 = iter2;
-    ASSERT_EQ((*iter).first, "foo");
+    ASSERT_EQ((*iter).get_oid(), "foo");
     ++iter;
-    ASSERT_TRUE(iter == ioctx.objects_end());
+    ASSERT_TRUE(iter == ioctx.nobjects_end());
     ++iter;
-    ASSERT_TRUE(iter == ioctx.objects_end());
+    ASSERT_TRUE(iter == ioctx.nobjects_end());
 
-    ASSERT_EQ(iter2->first, "foo");
-    ASSERT_EQ(iter3->first, "foo");
+    ASSERT_EQ(iter2->get_oid(), "foo");
+    ASSERT_EQ(iter3->get_oid(), "foo");
     ++iter2;
-    ASSERT_TRUE(iter2 == ioctx.objects_end());
+    ASSERT_TRUE(iter2 == ioctx.nobjects_end());
   }
 
-  ASSERT_EQ(iter3->first, "foo");
+  ASSERT_EQ(iter3->get_oid(), "foo");
   iter3 = iter3;
-  ASSERT_EQ(iter3->first, "foo");
+  ASSERT_EQ(iter3->get_oid(), "foo");
   ++iter3;
-  ASSERT_TRUE(iter3 == ioctx.objects_end());
+  ASSERT_TRUE(iter3 == ioctx.nobjects_end());
 }
 
 TEST_F(LibRadosListECPP, ListObjectsEndIter) {
@@ -463,26 +487,23 @@ TEST_F(LibRadosListECPP, ListObjectsEndIter) {
   bl1.append(buf, sizeof(buf));
   ASSERT_EQ(0, ioctx.write("foo", bl1, sizeof(buf), 0));
 
-  ObjectIterator iter(ioctx.objects_begin());
-  ObjectIterator iter_end(ioctx.objects_end());
-  ObjectIterator iter_end2 = ioctx.objects_end();
+  NObjectIterator iter(ioctx.nobjects_begin());
+  NObjectIterator iter_end(ioctx.nobjects_end());
+  NObjectIterator iter_end2 = ioctx.nobjects_end();
   ASSERT_TRUE(iter_end == iter_end2);
-  ASSERT_TRUE(iter_end == ioctx.objects_end());
-  ASSERT_TRUE(iter_end2 == ioctx.objects_end());
+  ASSERT_TRUE(iter_end == ioctx.nobjects_end());
+  ASSERT_TRUE(iter_end2 == ioctx.nobjects_end());
 
-  ASSERT_EQ(iter->first, "foo");
+  ASSERT_EQ(iter->get_oid(), "foo");
   ++iter;
-  ASSERT_TRUE(iter == ioctx.objects_end());
+  ASSERT_TRUE(iter == ioctx.nobjects_end());
   ASSERT_TRUE(iter == iter_end);
   ASSERT_TRUE(iter == iter_end2);
-  ObjectIterator iter2 = iter;
-  ASSERT_TRUE(iter2 == ioctx.objects_end());
+  NObjectIterator iter2 = iter;
+  ASSERT_TRUE(iter2 == ioctx.nobjects_end());
   ASSERT_TRUE(iter2 == iter_end);
   ASSERT_TRUE(iter2 == iter_end2);
 }
-
-#pragma GCC diagnostic pop
-#pragma GCC diagnostic warning "-Wpragmas"
 
 TEST_F(LibRadosListEC, ListObjectsNS) {
   char buf[128];
@@ -502,7 +523,7 @@ TEST_F(LibRadosListEC, ListObjectsNS) {
   ASSERT_EQ(0, rados_write(ioctx, "foo6", buf, sizeof(buf), 0));
   ASSERT_EQ(0, rados_write(ioctx, "foo7", buf, sizeof(buf), 0));
 
-  std::set<std::string> def, ns1, ns2;
+  std::set<std::string> def, ns1, ns2, all;
   def.insert(std::string("foo1"));
   def.insert(std::string("foo2"));
   def.insert(std::string("foo3"));
@@ -511,29 +532,39 @@ TEST_F(LibRadosListEC, ListObjectsNS) {
   ns1.insert(std::string("foo5"));
   ns2.insert(std::string("foo6"));
   ns2.insert(std::string("foo7"));
+  all.insert(std::string(":foo1"));
+  all.insert(std::string(":foo2"));
+  all.insert(std::string(":foo3"));
+  all.insert(std::string("ns1:foo1"));
+  all.insert(std::string("ns1:foo4"));
+  all.insert(std::string("ns1:foo5"));
+  all.insert(std::string("ns2:foo6"));
+  all.insert(std::string("ns2:foo7"));
 
   rados_list_ctx_t ctx;
   // Check default namespace ""
   rados_ioctx_set_namespace(ioctx, "");
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
-  check_list(def, ctx);
-  rados_objects_list_close(ctx);
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  check_list(def, ctx, "");
+  rados_nobjects_list_close(ctx);
 
   // Check default namespace "ns1"
   rados_ioctx_set_namespace(ioctx, "ns1");
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
-  check_list(ns1, ctx);
-  rados_objects_list_close(ctx);
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  check_list(ns1, ctx, "ns1");
+  rados_nobjects_list_close(ctx);
 
   // Check default namespace "ns2"
   rados_ioctx_set_namespace(ioctx, "ns2");
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
-  check_list(ns2, ctx);
-  rados_objects_list_close(ctx);
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  check_list(ns2, ctx, "ns2");
+  rados_nobjects_list_close(ctx);
 
-  // Can't specify all namespaces using old interface
+  // Check all namespaces
   rados_ioctx_set_namespace(ioctx, LIBRADOS_ALL_NSPACES);
-  ASSERT_EQ(-EINVAL, rados_objects_list_open(ioctx, &ctx));
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  check_list(all, ctx, all_nspaces);
+  rados_nobjects_list_close(ctx);
 }
 
 TEST_F(LibRadosListECPP, ListObjectsPPNS) {
@@ -567,21 +598,14 @@ TEST_F(LibRadosListECPP, ListObjectsPPNS) {
   ns2.insert(std::string("foo7"));
 
   ioctx.set_namespace("");
-  check_listpp(def, ioctx);
+  check_listpp(def, ioctx, "");
 
   ioctx.set_namespace("ns1");
-  check_listpp(ns1, ioctx);
+  check_listpp(ns1, ioctx, "ns1");
 
   ioctx.set_namespace("ns2");
-  check_listpp(ns2, ioctx);
-
-  ioctx.set_namespace(all_nspaces);
-  EXPECT_THROW(check_listpp(def, ioctx), std::runtime_error);
+  check_listpp(ns2, ioctx, "ns2");
 }
-
-#pragma GCC diagnostic ignored "-Wpragmas"
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
 
 TEST_F(LibRadosListECPP, ListObjectsManyPP) {
   char buf[128];
@@ -593,13 +617,13 @@ TEST_F(LibRadosListECPP, ListObjectsManyPP) {
     ASSERT_EQ(0, ioctx.write(stringify(i), bl, bl.length(), 0));
   }
 
-  librados::ObjectIterator it = ioctx.objects_begin();
+  librados::NObjectIterator it = ioctx.nobjects_begin();
   std::set<std::string> saw_obj;
   std::set<int> saw_pg;
-  for (; it != ioctx.objects_end(); ++it) {
-    std::cout << it->first
+  for (; it != ioctx.nobjects_end(); ++it) {
+    std::cout << it->get_oid()
 	      << " " << it.get_pg_hash_position() << std::endl;
-    saw_obj.insert(it->first);
+    saw_obj.insert(it->get_oid());
     saw_pg.insert(it.get_pg_hash_position());
   }
   std::cout << "saw " << saw_pg.size() << " pgs " << std::endl;
@@ -619,27 +643,27 @@ TEST_F(LibRadosListEC, ListObjectsStart) {
   }
 
   rados_list_ctx_t ctx;
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
   std::map<int, std::set<std::string> > pg_to_obj;
   const char *entry;
-  while (rados_objects_list_next(ctx, &entry, NULL) == 0) {
-    uint32_t pos = rados_objects_list_get_pg_hash_position(ctx);
+  while (rados_nobjects_list_next(ctx, &entry, NULL, NULL) == 0) {
+    uint32_t pos = rados_nobjects_list_get_pg_hash_position(ctx);
     std::cout << entry << " " << pos << std::endl;
     pg_to_obj[pos].insert(entry);
   }
-  rados_objects_list_close(ctx);
+  rados_nobjects_list_close(ctx);
 
   std::map<int, std::set<std::string> >::reverse_iterator p =
     pg_to_obj.rbegin();
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
   while (p != pg_to_obj.rend()) {
-    ASSERT_EQ((uint32_t)p->first, rados_objects_list_seek(ctx, p->first));
-    ASSERT_EQ(0, rados_objects_list_next(ctx, &entry, NULL));
+    ASSERT_EQ((uint32_t)p->first, rados_nobjects_list_seek(ctx, p->first));
+    ASSERT_EQ(0, rados_nobjects_list_next(ctx, &entry, NULL, NULL));
     std::cout << "have " << entry << " expect one of " << p->second << std::endl;
     ASSERT_TRUE(p->second.count(entry));
     ++p;
   }
-  rados_objects_list_close(ctx);
+  rados_nobjects_list_close(ctx);
 }
 
 TEST_F(LibRadosListECPP, ListObjectsStartPP) {
@@ -652,24 +676,91 @@ TEST_F(LibRadosListECPP, ListObjectsStartPP) {
     ASSERT_EQ(0, ioctx.write(stringify(i), bl, bl.length(), 0));
   }
 
-  librados::ObjectIterator it = ioctx.objects_begin();
+  librados::NObjectIterator it = ioctx.nobjects_begin();
   std::map<int, std::set<std::string> > pg_to_obj;
-  for (; it != ioctx.objects_end(); ++it) {
-    std::cout << it->first << " " << it.get_pg_hash_position() << std::endl;
-    pg_to_obj[it.get_pg_hash_position()].insert(it->first);
+  for (; it != ioctx.nobjects_end(); ++it) {
+    std::cout << it->get_oid() << " " << it.get_pg_hash_position() << std::endl;
+    pg_to_obj[it.get_pg_hash_position()].insert(it->get_oid());
   }
 
   std::map<int, std::set<std::string> >::reverse_iterator p =
     pg_to_obj.rbegin();
-  it = ioctx.objects_begin(p->first);
+  it = ioctx.nobjects_begin(p->first);
   while (p != pg_to_obj.rend()) {
     ASSERT_EQ((uint32_t)p->first, it.seek(p->first));
-    std::cout << "have " << it->first << " expect one of " << p->second << std::endl;
-    ASSERT_TRUE(p->second.count(it->first));
+    std::cout << "have " << it->get_oid() << " expect one of " << p->second << std::endl;
+    ASSERT_TRUE(p->second.count(it->get_oid()));
     ++p;
   }
 }
 
+TEST_F(LibRadosListPP, ListObjectsFilterPP) {
+  char buf[128];
+  memset(buf, 0xcc, sizeof(buf));
+  bufferlist obj_content;
+  obj_content.append(buf, sizeof(buf));
+
+  std::string target_str = "content";
+
+  // Write xattr bare, no ::encod'ing
+  bufferlist target_val;
+  target_val.append(target_str);
+  bufferlist nontarget_val;
+  nontarget_val.append("rhubarb");
+
+  ASSERT_EQ(0, ioctx.write("has_xattr", obj_content, obj_content.length(), 0));
+  ASSERT_EQ(0, ioctx.write("has_wrong_xattr", obj_content, obj_content.length(), 0));
+  ASSERT_EQ(0, ioctx.write("no_xattr", obj_content, obj_content.length(), 0));
+
+  ASSERT_EQ(0, ioctx.setxattr("has_xattr", "theattr", target_val));
+  ASSERT_EQ(0, ioctx.setxattr("has_wrong_xattr", "theattr", nontarget_val));
+
+  bufferlist filter_bl;
+  std::string filter_name = "plain";
+  ::encode(filter_name, filter_bl);
+  ::encode("_theattr", filter_bl);
+  ::encode(target_str, filter_bl);
+
+  NObjectIterator iter(ioctx.nobjects_begin(filter_bl));
+  bool foundit = false;
+  int k = 0;
+  while (iter != ioctx.nobjects_end()) {
+    foundit = true;
+    // We should only see the object that matches the filter
+    ASSERT_EQ((*iter).get_oid(), "has_xattr");
+    // We should only see it once
+    ASSERT_EQ(k, 0);
+    ++iter;
+    ++k;
+  }
+  ASSERT_TRUE(foundit);
+}
+
+TEST_F(LibRadosListNP, ListObjectsError) {
+  std::string pool_name;
+  rados_t cluster;
+  rados_ioctx_t ioctx;
+  pool_name = get_temp_pool_name();
+  ASSERT_EQ("", create_one_pool(pool_name, &cluster));
+  ASSERT_EQ(0, rados_ioctx_create(cluster, pool_name.c_str(), &ioctx));
+  char buf[128];
+  memset(buf, 0xcc, sizeof(buf));
+  rados_ioctx_set_namespace(ioctx, "");
+  ASSERT_EQ(0, rados_write(ioctx, "foo", buf, sizeof(buf), 0));
+  ASSERT_EQ(0, rados_pool_delete(cluster, pool_name.c_str()));
+
+  rados_list_ctx_t ctx;
+  ASSERT_EQ(0, rados_nobjects_list_open(ioctx, &ctx));
+  const char *entry;
+  ASSERT_EQ(-ENOENT, rados_nobjects_list_next(ctx, &entry, NULL, NULL));
+  rados_nobjects_list_close(ctx);
+  rados_ioctx_destroy(ioctx);
+  rados_shutdown(cluster);
+}
+
+
+
+// ---------------------------------------------
 
 TEST_F(LibRadosList, EnumerateObjects) {
   char buf[128];
@@ -886,27 +977,6 @@ TEST_F(LibRadosListPP, EnumerateObjectsSplitPP) {
   ASSERT_EQ(n_objects, saw_obj.size());
 }
 
-TEST_F(LibRadosListNP, ListObjectsError) {
-  std::string pool_name;
-  rados_t cluster;
-  rados_ioctx_t ioctx;
-  pool_name = get_temp_pool_name();
-  ASSERT_EQ("", create_one_pool(pool_name, &cluster));
-  ASSERT_EQ(0, rados_ioctx_create(cluster, pool_name.c_str(), &ioctx));
-  char buf[128];
-  memset(buf, 0xcc, sizeof(buf));
-  rados_ioctx_set_namespace(ioctx, "");
-  ASSERT_EQ(0, rados_write(ioctx, "foo", buf, sizeof(buf), 0));
-  ASSERT_EQ(0, rados_pool_delete(cluster, pool_name.c_str()));
-  
-  rados_list_ctx_t ctx;
-  ASSERT_EQ(0, rados_objects_list_open(ioctx, &ctx));
-  const char *entry;
-  ASSERT_EQ(-ENOENT, rados_objects_list_next(ctx, &entry, NULL));
-  rados_objects_list_close(ctx);
-  rados_ioctx_destroy(ioctx);
-  rados_shutdown(cluster);
-}
 
 TEST_F(LibRadosListPP, EnumerateObjectsFilterPP) {
   char buf[128];
